moved armor page helper files
