+++ src/app/armor/ArmorResultRow.tsx
