+++ src/app/armor/InputNumber.tsx
