+++ src/app/armor/InputSelect.tsx
